Split up LogicalBatchReader/Writer
